PFBlockProducer to one file
