--- RecoParticleFlow/PFProducer/plugins/PFBlockProducer.h
@@ -1,72 +1,0 @@
-#ifndef RecoParticleFlow_PFProducer_PFBlockProducer_h_
-#define RecoParticleFlow_PFProducer_PFBlockProducer_h_
-
-// system include files
-#include <memory>
-#include <string>
-
-// user include files
-#include "FWCore/Framework/interface/Frameworkfwd.h"
-#include "FWCore/Framework/interface/stream/EDProducer.h"
-#include "FWCore/ParameterSet/interface/ParameterSet.h"
-
-#include "FWCore/Framework/interface/Event.h"
-#include "FWCore/Framework/interface/MakerMacros.h"
-#include "FWCore/Utilities/interface/EDPutToken.h"
-
-#include "RecoParticleFlow/PFProducer/interface/PFBlockAlgo.h"
-
-#include "DataFormats/ParticleFlowReco/interface/PFLayer.h"
-#include "DataFormats/ParticleFlowReco/interface/PFCluster.h"
-#include "DataFormats/ParticleFlowReco/interface/PFRecTrack.h" 
-#include "DataFormats/ParticleFlowReco/interface/PFDisplacedVertexFwd.h"
-#include "DataFormats/ParticleFlowReco/interface/PFDisplacedVertex.h"
-#include "DataFormats/ParticleFlowReco/interface/PFConversionFwd.h"
-#include "DataFormats/ParticleFlowReco/interface/PFConversion.h"
-#include "DataFormats/ParticleFlowReco/interface/PFV0Fwd.h"
-#include "DataFormats/ParticleFlowReco/interface/PFV0.h"
-
-#include "DataFormats/EgammaCandidates/interface/Photon.h"
-#include "DataFormats/EgammaCandidates/interface/PhotonFwd.h"
-
-
-/**\class PFBlockProducer 
-\brief Producer for particle flow blocks
-
-This producer makes use of PFBlockAlgo, the particle flow block algorithm.
-Particle flow itself consists in reconstructing particles from the particle 
-flow blocks This is done at a later stage, see PFProducer and PFAlgo.
-
-\author Colin Bernet
-\date   April 2007
-*/
-
-class FSimEvent;
-
-
-
-class PFBlockProducer : public edm::stream::EDProducer<> {
- public:
-
-  explicit PFBlockProducer(const edm::ParameterSet&);
-
-  ~PFBlockProducer() override;
-  
-  void beginLuminosityBlock(edm::LuminosityBlock const&, 
-				    edm::EventSetup const&) override;
-
-  void produce(edm::Event&, const edm::EventSetup&) override;
-
- private:
-  /// verbose ?
-  const bool   verbose_;
-  const edm::EDPutTokenT<reco::PFBlockCollection> putToken_;
-  
-  /// Particle flow block algorithm 
-  PFBlockAlgo            pfBlockAlgo_;
-
-};
-
-DEFINE_FWK_MODULE(PFBlockProducer);
-
-#endif
